moves stream closing test to a separate class
